In the first version there was no check for the number of characters in a line, so if the number of characters exceeded MAX_ROW_CHARS you could never notice. In this version, this is corrected and if these two come close the program will abort and notify you.
In the previous version the number of rows and number of columns were different macros but now they are one: BUFFER_NUM. In case the data has more comments or more rows the program will add a unit of BUFFER_NUM elements to them respectively. It will no longer give an error in the case of comments.

Also the "relloc"s in charge of defining and possibly extending the data and comment arrays have now gone into their respective functions making the readasciitable function much easier to read.